Generate yaml files for discovered resources
Instead of having a hard coded listed of resources that use to generate
yaml files, we now attempt to discover all the resources for each node
specified when launching the support-bundle-kit.

This removed the explicit handling of harvester resources and the
external longhorn bundle.  The explicit kubernetes yaml file generate
remains. It might be redundant with the discovery of resources.

Signed-off-by: Keith Lucas <keith.lucas@suse.com>